pyproject.toml
[build-system]
requires      = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["traenslenzor*"]
exclude = ["data*"]

[project]
name = "traenslenzor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=23.2.1",
    "fastapi>=0.120.4",
    "fastmcp>=2.12.4",
    "mcp>=1.0.0",
    "matplotlib>=3.10.7",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "httpx>=0.27.0",
    "langchain>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.0",
    "numpy>=1.26.4,<2.3.0",
    "opencv-python>=4.11.0.86",
    "types-aiofiles>=25.1.0.20251011",
    "requests>=2.31.0",
    "transformers>=4.30.0",
    "uvicorn>=0.37.0",
    "ollama>=0.6.0",
    "streamlit>=1.52.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pandas-stubs>=2.3.2.250926",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-image-diff>=0.0.14",
    "ruff>=0.14.1",
    "types-defusedxml>=0.7.0.20250822",
    "types-pexpect>=4.9.0.20250916",
    "types-psutil>=7.0.0.20251001",
    "types-pygments>=2.19.0.20250809",
    "types-pysocks>=1.7.1.20251001",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
    "albumentations>=2.0.8",
    "datasets>=4.4.1",
    "devtools>=0.12.2",
    "tomli-w>=1.0.0",
    "optuna>=4.2.0",
    "optuna-integration[pytorch-lightning]>=4.5.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "lightning>=2.5.6",
    "rich>=13.9.4",
    "torchmetrics>=1.6.0",
    "torchsummary>=1.5.1",
    "captum>=0.8.0",
    "wandb>=0.22.3",
    "jaxtyping>=0.3.3",
    "typing-extensions>=4.15.0",
    "syrenka>=0.5.7",
    "watchdog>=6.0.0",
    "py-reform>=0.1.0",
    "openvino>=2025.4.1",
    "pytesseract>=0.3.13",
]

notebook = [
    "ipympl>=0.9.0",
    "jupyter>=1.0.0",
    "ipykernel>=7.1.0",
    "notebook>=7.0.0",
    "jupyterlab>=4.5.0",
]


[tool.pytest.ini_options]
testpaths  = ["tests"]
pythonpath = ["traenslenzor"]
addopts    = "-q"

[tool.ruff]
line-length    = 100
target-version = "py313"

    [tool.ruff.lint]
    extend-select = ["I"]    # import sorting
    ignore        = ["F722"] # allow jaxtyping

        [tool.ruff.lint.per-file-ignores]
        "tests/*" = ["S101"]
